stream.segmented: move check_queue_deadline()
- Move `check_queue_deadline()` to `SegmentedStreamWorker` base class
- Add `_queue_deadline_wait` property for subclasses to override
- Add queue deadline check to `DASHStreamWorker.iter_segments()`
  for dynamic manifests with `SegmentList` representations (polling)
